refactor(lotto): domain 안에서 패키지 분리
