+++ src/main/kotlin/lotto/domain/Lotto.kt
@@ -1,5 +1,8 @@
 package lotto.domain
 
+import lotto.domain.generator.AutoLottoGenerator
+import lotto.domain.generator.LottoGenerator
+
 data class Lotto(private val generator: LottoGenerator = AutoLottoGenerator) {
     val numbers = generator.execute().sorted()
