+++ src/main/kotlin/lotto/domain/generator/LottoGenerator.kt
@@ -1,4 +1,4 @@
-package lotto.domain
+package lotto.domain.generator
 
 interface LottoGenerator {
