+++ src/main/kotlin/lotto/domain/selling/Payment.kt
@@ -1,4 +1,6 @@
-package lotto.domain
+package lotto.domain.selling
+
+import lotto.domain.LottoType
 
 data class Payment(
     val money: Int,
