+++ src/test/kotlin/lotto/domain/generator/ManualLottoGeneratorTest.kt
@@ -1,4 +1,4 @@
-package lotto.domain
+package lotto.domain.generator
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.DisplayName
